pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qualer-sdk"
version = "2.2.1"
description = "Auto-generated Python client for Qualer API"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    {name = "Jeff Hall", email = "jhall@jgiquality.com"}
]
keywords = ["qualer", "api", "sdk", "client"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
    "Typing :: Typed",
]
dependencies = [
    "urllib3>=1.15",
    "six>=1.10",
    "certifi",
    "python-dateutil",
    "pydantic>=2.11.0,<2.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "mypy",
    "black",
    "flake8",
    "isort",
    "autoflake",
]

[project.urls]
Homepage = "https://github.com/Johnson-Gage-Inspection-Inc/qualer-sdk-python"
Repository = "https://github.com/Johnson-Gage-Inspection-Inc/qualer-sdk-python"
Issues = "https://github.com/Johnson-Gage-Inspection-Inc/qualer-sdk-python/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
qualer_sdk = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py37']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
disable_error_code = ["no-any-return"]
exclude = [
    "build",
    "dist", 
    ".*\\.egg-info",
    "__pycache__",
    "\\.pytest_cache",
    "temp_sdk_gen",
    "src_old"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/qualer_sdk"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/build/*",
    "*/dist/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]
